Reviewing nodejs resource server example
