+++ nodejs/app-nodejs-html5/src/main/webapp/keycloak.json
@@ -1,0 +1,7 @@
+{
+  "realm": "quickstart",
+  "auth-server-url": "http://localhost:8180/auth",
+  "ssl-required": "external",
+  "resource": "app-html5",
+  "public-client": true
+}
