--- nodejs/service-nodejs/client-import.json
@@ -1,5 +1,0 @@
-{
-    "clientId": "service-nodejs",
-    "enabled": true,
-    "bearerOnly": true
-}
