+++ nodejs/service-nodejs/test/utils/config.js
@@ -1,0 +1,20 @@
+const baseUrl = 'http://127.0.0.1:8180/auth';
+
+export default {
+  baseUrl: baseUrl,
+  token: {
+    username: 'alice',
+    password: 'password',
+    grant_type: 'password',
+    client_id: 'test-cli',
+    realmName: 'quickstart'
+  },
+  adminClient: {
+    baseUrl: baseUrl,
+    realmName: 'master',
+    username: 'admin',
+    password: 'admin',
+    grantType: 'password',
+    clientId: 'admin-cli'
+  }
+};
