+++ nodejs/service-nodejs/test/utils/delete-realm.js
@@ -1,0 +1,3 @@
+import adminClient from './keycloak-admin-client.js';
+
+await adminClient.realms.del({ realm : 'quickstart' });
