+++ scripts/start-local-server.sh
@@ -11,4 +11,4 @@ fi
 
 keycloak-dist/bin/kc.sh start-dev --http-port=8180 --http-relative-path="/auth" $SERVER_ARGS > keycloak.log 2>&1 &
 
-wget --retry-connrefused --waitretry=1 --read-timeout=20 --timeout=15 -t 30 http://localhost:8180/auth
+wget --retry-connrefused --waitretry=3 --read-timeout=20 --timeout=15 -t 30 http://localhost:8180/auth
